[audio] do some clean up
