+++ packages/expo-audio/ios/Utils.swift
@@ -1,0 +1,12 @@
+import AVFoundation
+
+func createAVPlayer(source: AudioSource?) -> AVPlayer {
+  if let source, let url = source.uri {
+    do {
+      return try AVPlayer(url: url)
+    } catch {
+      return AVPlayer()
+    }
+  }
+  return AVPlayer()
+}
